requirements-phase1.txt
# Phase 1 核心依赖
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.34.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# 可选依赖（如果需要更多功能）
# python-docx>=1.1.0
# pandas>=2.1.3